Crate runtime_raw

Types for defining custom Runtimes. See the Runtime documentation for more details.

These types are only necessary when implementing custom runtimes. If you're only trying to perform IO, then there's no need to bother with any of these types as they will have been implemented for you already.

Traits

Runtime
The runtime trait.

TcpListener
A TcpListener for this Runtime

TcpStream
A TcpStream for this Runtime

UdpSocket
A UDP socket.

Functions

current_runtime
Get the current runtime.

enter
Runs a future inside a runtime and blocks on the result.

set_runtime
Set the current runtime.